Replace flex layout with grid layout.
Neither Grid nor Flex layouts seem to properly support page breaks in
print media in Chrome. To get past this, I needed to fiddle with the
height to add and remove arbitrary amounts.

This may require further fiddling to keep the layout working as expected
with a large number of pages.

Most measurements are based on US Letter paper size.